Price type to decimal to match api

+++ ThAmCo.Staff/Models/OrderDetail.cs
@@ -1,11 +1,9 @@
-﻿namespace ThAmCo.Staff.Models
-{
-    public class OrderDetail
-    {
+﻿namespace ThAmCo.Staff.Models {
+    public class OrderDetail {
         public int Id { get; set; }
         public int OrderId { get; set; }
         public int ProductId { get; set; }
-        public double UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
         public int Quantity { get; set; }
     }
 }
